enem-analytics/backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
pydantic[email]
python-multipart
xgboost
joblib
scipy
requests
# Authentication
python-jose[cryptography]
passlib[bcrypt]
# Database
sqlalchemy
psycopg2-binary
python-dotenv
